pyproject.toml
[tool.poetry]
name = "cp-abe-based-on-ecc-implementation"
version = "0.1.0"
description = ""
authors = ["hggforget <1242537651@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-socketio = {extras = ["client"], version = "^5.9.0"}
aiohttp = "^3.8.5"
nummaster = "^0.1.1"
requests = "^2.31.0"
websocket-client = "^1.6.3"
eventlet = "^0.33.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
